refactor job data wrangling
